Props, Pickups, Enemy Spawner
Breakable props are now in the game with a chance to spawn some of the pickups.
Drop Chance Manager is edited so now it only spawns 1 of the possible items instead of spawning multiple if the chances were right.
There are now 3 different gem pickups that all grand different amount of experience.
There's 2 potions that heal a different amount (small, big)
Magnet pickup that increases the magnet stat of the player to attract all pickups on the screen.
Speed Boost Boots that increase the player's movement speed when picked up.
Enemy Spawner is now improved so waves can end 2 ways. Either the timer runs out and it automatically spawns the next wave, or the player kills all minions and the new wave is spawned.