LevelUpFX, DropRateManager, balance changes
A level up animation and a short delay is added to the game so players don't accidentally spend their reward point without even looking at what weapon they are upgrading.

DropRateManager is improved so now the dropping doesn't happen onDestroy, but rather on the call of the spawn function. This way when an enemy is Killed the drop is instantiated immediately instead of having to wait for the dying animation to finish.

Some slight balance changes, trying to make the game feel a bit more fluid and fast phased