Dash, Particles, Trail
Player can now dash a short distance. While dashing the collision with enemies is turned off and the player is invincible.

Player is drawing a short trail effect after themselves to make it more fancy when dashing.

Explosion effect is upgraded to particle  system instead of just simply playing an explosion animation.